requirements.txt
flask==3.1.1
requests==2.32.4
matplotlib==3.10.5
python-dotenv==1.1.1
gunicorn==21.2.0

